\name{boxplot.microbenchmark}
\alias{boxplot.microbenchmark}
\title{Boxplot of microbenchmark timings.}
\usage{\method{boxplot}{microbenchmark}(x, unit="t", log=TRUE, xlab, ylab, ...)
}
\description{Boxplot of \code{microbenchmark} timings.}
\author{Olaf Mersmann \email{olafm@datensplitter.net}}
\arguments{\item{x}{A \code{microbenchmark} object.}
\item{unit}{Unit in which the results be plotted.}
\item{log}{Should times be plotted on log scale?}
\item{xlab}{X axes label.}
\item{ylab}{Y axes label.}
\item{...}{Passed on to boxplot.formula.}
}

