\name{coalesce}
\alias{coalesce}
\title{Return first non null argument.}
\usage{coalesce(...)
}
\description{Return first non null argument.}
\details{This function is useful when processing complex arguments with multiple
possible defaults based on other arguments that may or may not have been
provided.}
\value{First non null element in \code{...}.}
\author{Olaf Mersmann \email{olafm@statistik.tu-dortmund.de}}
\arguments{\item{...}{List of values.}
}

