% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanotime.R
\name{get_nanotime}
\alias{get_nanotime}
\title{Return the current value of the platform timer.}
\usage{
get_nanotime()
}
\description{
The current value of the most accurate timer of the platform is
returned. This can be used as a time stamp for logging or similar
purposes. Please note that there is no common reference, that is,
the timer value cannot be converted to a date and time value.
}
\author{
Olaf Mersmann
}
