\name{ggplot.microbenchmark}
\alias{ggplot.microbenchmark}
\title{ggplot method for microbenchmark objects.}
\usage{\method{ggplot}{microbenchmark}(data, mapping=aes_string(x = "run", y = "value", colour = "expr"), ...)
}
\description{ggplot method for \code{microbenchmark} objects.}
\author{Olaf Mersmann \email{olafm@datensplitter.net}}
\arguments{\item{data}{A \code{microbenchmark} object.}
\item{mapping}{Default list of aesthetic mappings.}
\item{...}{Passed to \code{\link{ggplot.data.frame}}.}
}

