% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{normalize_unit}
\alias{normalize_unit}
\title{Normalize timing units to one of the supported values}
\usage{
normalize_unit(unit)
}
\arguments{
\item{unit}{A unit of time. See details.}
}
\value{
A matrix containing the converted time values with an
  attribute \code{unit} which is a printable name of the unit of
  time.
}
\description{
We support the following units of time
\describe{
  \item{\dQuote{ns}, \dQuote{nanoseconds}}{}
  \item{\dQuote{us}, \dQuote{microseconds}}{}
  \item{\dQuote{ms}, \dQuote{milliseconds}}{}
  \item{\dQuote{s}, \dQuote{secs}, \dQuote{seconds}}{}
  \item{\dQuote{t}, \dQuote{time}}{Appropriately prefixed time unit.}
  \item{\dQuote{eps}}{Evaluations per second / Hertz.}
  \item{\dQuote{hz}}{Hertz / evaluations per second.}
  \item{\dQuote{khz}}{Kilohertz / 1000s of evaluations per second.}
  \item{\dQuote{mhz}}{Megahertz / 1000000s of evaluations per second.}
  \item{\dQuote{f}, \dQuote{frequency}}{Appropriately prefixed frequency unit.}
}
}
\author{
Joshua M. Ulrich
}
\keyword{internal}
