% Generated by roxygen2: do not edit by hand
\name{system.nanotime}
\alias{system.nanotime}
\title{CPU time used measured in sub-milliseconds}
\usage{
system.nanotime(expr, gcFirst = TRUE)
}
\arguments{
\item{expr}{Valid R expression to be timed.}

\item{gcFirst}{Logical - should a garbage collection be performed immediately before the timing? Default is TRUE.}
}
\value{
\code{proc_time} class object, only \code{elapsed} value is returned, all others are \code{NA}.
}
\description{
Clone of base R \code{system.time} function utilizing accuracy of \code{microbenchmark} package.
}
\seealso{
\code{\link{get_nanotime}}
}

