% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.microbenchmark}
\alias{autoplot.microbenchmark}
\title{Autoplot method for microbenchmark objects: Prettier graphs for
microbenchmark using ggplot2}
\usage{
\method{autoplot}{microbenchmark}(
  object,
  expression,
  breaks = c(0, 1e-05, 1e-04, 0.001, 0.01, 0.1, 1, 10, 100, 1000, 10000, 1e+05),
  ...,
  log = TRUE,
  y_min = 0.95 * min(object$time),
  y_max = 1.05 * max(object$time)
)
}
\arguments{
\item{object}{A microbenchmark object}

\item{expression}{Label of valued expression}

\item{breaks}{break of scale_y_log10 if \code{log=TRUE}.}

\item{\dots}{hyperparameter of  scale_y_continuous}

\item{log}{If \code{TRUE} the time axis will be on log scale.}

\item{y_min}{The lower limit of the y axis (defaults to 5 percent less than
the minimum value)}

\item{y_max}{The upper limit of the y axis (defaults to 5 percent more than
the maximum value)}
}
\value{
A ggplot2 plot
}
\description{
Uses ggplot2 to produce a more legible graph of microbenchmark timings
}
\author{
Ari Friedman, Olaf Mersmann, Florian Robert
}
