% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.microbenchmark}
\alias{autoplot.microbenchmark}
\title{Autoplot method for microbenchmark objects: Prettier graphs for
microbenchmark using ggplot2}
\usage{
autoplot.microbenchmark(object, unit, ..., log = TRUE, y_max = NULL)
}
\arguments{
\item{object}{A microbenchmark object}

\item{unit}{What unit to use on the time axis. Defaults to the \code{unit}
attribute of \code{object}, or the option \code{microbenchmark.unit} if the
attribute is not set. If neither is set, \dQuote{t} is used.}

\item{\dots}{Ignored}

\item{log}{If \code{TRUE} the time axis will be on log scale.}

\item{y_max}{The maximum time to plot in \code{unit}s (nanoseconds if the
unit is not given, or is \dQuote{t} or \dQuote{f}). Note that if
\code{unit} is a frequency, the maximum time corresponds to the minimum
frequency on the axis.}
}
\value{
A ggplot2 plot
}
\description{
Uses ggplot2 to produce a more legible graph of microbenchmark timings
}
\examples{
if (requireNamespace("ggplot2")) {

tm <- microbenchmark(rchisq(100, 0),
                     rchisq(100, 1),
                     rchisq(100, 2),
                     rchisq(100, 3),
                     rchisq(100, 5), times=1000L)
ggplot2::autoplot(tm)
}
}
\author{
Ari Friedman, Olaf Mersmann, Thomas Nygreen
}
