% Generated by roxygen2 (4.0.2): do not edit by hand
\name{boxplot.microbenchmark}
\alias{boxplot.microbenchmark}
\title{Boxplot of \code{microbenchmark} timings.}
\usage{
\method{boxplot}{microbenchmark}(x, unit = "t", log = TRUE, xlab, ylab, ...)
}
\arguments{
\item{x}{A \code{microbenchmark} object.}

\item{unit}{Unit in which the results be plotted.}

\item{log}{Should times be plotted on log scale?}

\item{xlab}{X axes label.}

\item{ylab}{Y axes label.}

\item{...}{Passed on to boxplot.formula.}
}
\description{
Boxplot of \code{microbenchmark} timings.
}
\author{
Olaf Mersmann \email{olafm@p-value.net}
}

