% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot.microbenchmark}
\alias{boxplot.microbenchmark}
\title{Boxplot of \code{microbenchmark} timings.}
\usage{
\method{boxplot}{microbenchmark}(
  x,
  unit = "t",
  log = TRUE,
  xlab,
  ylab,
  horizontal = FALSE,
  main = "microbenchmark timings",
  ...
)
}
\arguments{
\item{x}{A \code{microbenchmark} object.}

\item{unit}{Unit in which the results be plotted.}

\item{log}{Should times be plotted on log scale?}

\item{xlab}{X axis label.}

\item{ylab}{Y axis label.}

\item{horizontal}{Switch X and Y axes.}

\item{main}{Plot title.}

\item{...}{Passed on to boxplot.formula.}
}
\description{
Boxplot of \code{microbenchmark} timings.
}
\author{
Olaf Mersmann
}
