% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{coalesce}
\alias{coalesce}
\title{Return first non null argument.}
\usage{
coalesce(...)
}
\arguments{
\item{...}{List of values.}
}
\value{
First non null element in \code{...}.
}
\description{
This function is useful when processing complex arguments with multiple
possible defaults based on other arguments that may or may not have been
provided.
}
\author{
Olaf Mersmann
}

