% Generated by roxygen2 (4.0.2): do not edit by hand
\name{microtiming_precision}
\alias{microtiming_precision}
\title{Estimate precision of timing routines.}
\usage{
microtiming_precision(rounds = 100L, warmup = 2^18)
}
\arguments{
\item{rounds}{Number of measurements used to estimate the precision.}

\item{warmup}{Number of iterations used to warm up the CPU.}
}
\value{
A vector of observed non-zero timings.
}
\description{
This function is currently experimental. Its main use is to judge
the quality of the underlying timer implementation of the
operating system. The function measures the overhead of timing a C
function call \code{rounds} times and returns all non-zero timings
observed. This can be used to judge the granularity and resolution
of the timing subsystem.
}
\author{
Olaf Mersmann \email{olafm@statistik.tu-dortmund.de}
}

