% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.microbenchmark}
\alias{summary.microbenchmark}
\title{Summarize \code{microbenchmark} timings.}
\usage{
\method{summary}{microbenchmark}(object, unit, ..., include_cld = TRUE)
}
\arguments{
\item{object}{An object of class \code{microbenchmark}.}

\item{unit}{What unit to print the timings in. If none is given,
either the \code{unit} attribute of \code{object} or the option
\code{microbenchmark.unit} is used and if neither is set
\dQuote{t} is used.}

\item{...}{Passed to \code{print.data.frame}}

\item{include_cld}{Calculate \code{cld} using \code{multcomp::glht()}
and add it to the output. Set to \code{FALSE} if the calculation takes
too long.}
}
\value{
A \code{data.frame} containing the aggregated results.
}
\description{
Summarize \code{microbenchmark} timings.
}
\note{
The available units are nanoseconds (\code{"ns"}),
microseconds (\code{"us"}), milliseconds (\code{"ms"}), seconds
(\code{"s"}) and evaluations per seconds (\code{"eps"}) and
relative runtime compared to the best median time
(\code{"relative"}).
}
\seealso{
\code{\link{print.microbenchmark}}
}
