% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyplot.microbenchmark.R
\name{tinyplot.microbenchmark}
\alias{tinyplot.microbenchmark}
\title{Tinyplot method for microbenchmark objects}
\usage{
\method{tinyplot}{microbenchmark}(
  x,
  type = c("violin", "boxplot", "jitter"),
  log = TRUE,
  unit = NULL,
  order = NULL,
  main = "microbenchmark timings",
  xlab = NA,
  ylab = NULL,
  flip = TRUE,
  trim = TRUE,
  joint.bw = c("none", "mean", "full"),
  ...
)
}
\arguments{
\item{x}{A microbenchmark object.}

\item{type}{String giving the type of plot representation. One of `"violin"`
(the default), `"boxplot"`, or `"jitter"`.}

\item{log}{Should times be plotted on a log scale? Default is \code{TRUE}.}

\item{unit}{Unit in which the results be plotted.}

\item{order}{Names of output column(s) to order the results.}

\item{main, xlab, ylab}{Plot and axes titles.}

\item{flip}{Switch the X and Y axes? Default is \code{TRUE}.}

\item{trim}{Trim violin plots to data extent? Default is \code{TRUE}.}

\item{joint.bw}{Which (if any) joint smoothing bandwidth to use on violin
plots? Default is \code{"none"} to match
\code{\link[microbenchmark]{autoplot.microbenchmark}}.}

\item{...}{Additional arguments passed to [`tinyplot`].}
}
\value{
No return value. Called for side effect of producing a plot.
}
\description{
Uses the `tinyplot` package to produce prettier base graphics
  for microbenchmark timings. Note that `tinyplot` needs to be installed and
  loaded separately.
}
\examples{
if (requireNamespace("tinyplot", quietly = TRUE)) {
    library(tinyplot)
    
    tm <- microbenchmark(rchisq(100, 0),
                         rchisq(100, 1),
                         rchisq(100, 2),
                         rchisq(100, 3),
                         rchisq(100, 5), times=1000L)
    
    tinyplot(tm)
    
    # aesthetic tweak example with custom title
    tinytheme("classic")
    tinyplot(tm, fill = "transparent",
             main = "Impressive benchmarks",
             sub = "Brought to you by tinyplot")

    # we can use the tinyplot scaffolding to add layers to our plot
    tinyplot_add(type = "jitter", pch = ".", alpha = 0.3)
    
    # reset theme
    tinytheme()
}
}
\author{
Grant McDermott
}
