% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyplot.microbenchmark.R
\name{tinyplot.microbenchmark}
\alias{tinyplot.microbenchmark}
\title{Tinyplot method for microbenchmark objects}
\usage{
\method{tinyplot}{microbenchmark}(
  x,
  type = c("violin", "boxplot", "jitter"),
  order = NULL,
  log = TRUE,
  unit = NULL,
  main = "microbenchmark timings",
  xlab = NA,
  ylab = NULL,
  ylim = NULL,
  flip = TRUE,
  trim = TRUE,
  joint.bw = FALSE,
  ...
)
}
\arguments{
\item{x}{A microbenchmark object.}

\item{type}{String giving the type of plot representation. One of `"violin"`
(the default), `"boxplot"`, or `"jitter"`.}

\item{order}{Names of output column(s) to order the results.}

\item{log}{If \code{TRUE} the time axis will be on log scale.}

\item{unit}{The unit to use for graph labels.}

\item{main, xlab, ylab}{Main and axes titles of the plot.}

\item{ylim}{Numeric vector of length 2, giving the limits of the timings in 
the unit automatically chosen for for the time axis. (Note that, in the
default case where `flip = TRUE`, the timings will appear on the x-axis.)
If no argument is provided, then will revert to the minimum and maximum
recorded values.}

\item{flip}{If \code{TRUE} the plot will be flipped on its side (default).}

\item{trim}{If \code{TRUE} the violin plots will be trimmed.}

\item{joint.bw}{If \code{TRUE} use a joint bandwidth for violin plots.}

\item{...}{Additional arguments passed to [`tinyplot`].}
}
\value{
No return value. Called for side effect of producing a plot.
}
\description{
Uses the `tinyplot` package to produce prettier base graphics
  for microbenchmark timings. Note that `tinyplot` needs to be installed and
  loaded separately.
}
\examples{
if (requireNamespace("tinyplot", quietly = TRUE)) {
    library(tinyplot)
    
    tm <- microbenchmark(rchisq(100, 0),
                         rchisq(100, 1),
                         rchisq(100, 2),
                         rchisq(100, 3),
                         rchisq(100, 5), times=1000L)
    
    tinyplot(tm)
    
    # aesthetic tweak example with custom title
    tinytheme("classic")
    tinyplot(tm, fill = "transparent",
             main = "Impressive benchmarks",
             sub = "Brought to you by tinyplot")

    # we can use the tinyplot scaffolding to add layers to our plot
    tinyplot_add(type = "jitter", pch = ".", alpha = 0.3)
    
    # reset theme
    tinytheme()
}
}
\author{
Grant McDermott
}
